uwp: add support for arm64
